refactor(LottoRules): 공통된 상수를 제공하는 클래스 생성
- Lotto와 LottoMachine에서 사용하는 공통 상수를 클래스로 분리
