+++ src/main/java/lotto/constants/LottoRules.java
@@ -1,0 +1,8 @@
+package lotto.constants;
+
+public class LottoRules {
+    private LottoRules(){}
+    public static final int LOTTO_MIN_NUMBER = 1;
+    public static final int LOTTO_MAX_NUMBER = 45;
+    public static final int LOTTO_COUNT = 6;
+}
