+++ src/main/java/lotto/infra/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.infra;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
